Enable GitHubApi to manage an app's JWT
This change makes it possible to cache the app ID and private key for a
GitHub App in the GitHubApi client. When set, the client will
automatically generate and use a JWT to authenticate as the GitHub App
(similar to caching the oauth_token on GitHubApi when working as an app
installation or with a PAT). The cached JWT is available from the `jwt`
property on the class. When the `jwt` property is accessed, the cached
jwt token's expiration is tested, and a new jwt is automatically
generated if necessary.